api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
firebase-admin>=6.5.0
pydantic>=2.9.0
python-dotenv>=1.0.1
httpx>=0.27.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
python-multipart>=0.0.9
mcp>=1.26.0
cryptography>=42.0.0
apscheduler>=3.10.0
